+++ announcements/src/org/labkey/announcements/AnnouncementsController.java
@@ -26,6 +26,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.json.JSONObject;
 import org.labkey.announcements.model.AnnouncementDigestProvider;
+import org.labkey.announcements.model.AnnouncementFullModel;
 import org.labkey.announcements.model.AnnouncementManager;
 import org.labkey.announcements.model.AnnouncementModel;
 import org.labkey.announcements.model.DailyDigestEmailPrefsSelector;
@@ -938,7 +939,7 @@ public class AnnouncementsController extends SpringActionController
         public ModelAndView getInsertUpdateView(AnnouncementForm form, boolean reshow, BindException errors)
         {
             Permissions perm = getPermissions();
-            AnnouncementModel parent = null;
+            AnnouncementFullModel parent = null;
             Container c = getContainer();
 
             if (null != form.getParentId())
@@ -2306,7 +2307,7 @@ public class AnnouncementsController extends SpringActionController
 
     public static class ThreadViewBean
     {
-        public AnnouncementModel announcementModel;
+        public AnnouncementFullModel announcementModel;
         public String message = "";
         public Permissions perm = null;
         public boolean isResponse = false;
@@ -2334,7 +2335,7 @@ public class AnnouncementsController extends SpringActionController
             init(c, findThread(c, rowId, entityId), currentURL, getPermissions(c, user, getSettings(c)), false, false);
         }
 
-        public ThreadView(Container c, ActionURL url, AnnouncementModel ann, Permissions perm)
+        public ThreadView(Container c, ActionURL url, AnnouncementFullModel ann, Permissions perm)
         {
             this();
             init(c, ann, url, perm, true, false);
@@ -2343,11 +2344,11 @@ public class AnnouncementsController extends SpringActionController
         public ThreadView(AnnouncementForm form, Container c, ActionURL url, Permissions perm, boolean print)
         {
             this();
-            AnnouncementModel ann = findThread(c, form.get("rowId"), form.get("entityId"));
+            AnnouncementFullModel ann = findThread(c, form.get("rowId"), form.get("entityId"));
             init(c, ann, url, perm, false, print);
         }
 
-        protected void init(Container c, AnnouncementModel ann, URLHelper currentURL, Permissions perm, boolean isResponse, boolean print)
+        protected void init(Container c, AnnouncementFullModel ann, URLHelper currentURL, Permissions perm, boolean isResponse, boolean print)
         {
             if (null == c || !perm.allowRead(ann))
             {
@@ -2452,7 +2453,7 @@ public class AnnouncementsController extends SpringActionController
     }
 
 
-    private static @Nullable AnnouncementModel findThread(Container c, String rowIdVal, String entityId)
+    private static @Nullable AnnouncementFullModel findThread(Container c, String rowIdVal, String entityId)
     {
         int rowId = 0;
         if (rowIdVal != null)
